visual_embeddings/requirements.txt
filelock==3.13.3
fsspec==2024.3.1
ftfy==6.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
pillow==10.2.0
regex==2023.12.25
sympy==1.12
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.10.0
wcwidth==0.2.13
